+++ apps/rush/src/utilities/Telemetry.ts
@@ -20,6 +20,8 @@ export interface ITelemetryData {
   extraData?: { [key: string]: string };
 }
 
+const MAX_FILE_COUNT: number = 100;
+
 export default class Telemetry {
   private _enabled: boolean;
   private _store: ITelemetryData[];
@@ -45,17 +47,53 @@ export default class Telemetry {
   }
 
   public flush(writeFile: (file: string, data: string) => void = fsx.writeFileSync): void {
-    if (!this._enabled) {
+    if (!this._enabled || this._store.length === 0) {
       return;
     }
+
     const fullPath: string = this._getFilePath();
     fsx.ensureDirSync(this._dataFolder);
     writeFile(fullPath, JSON.stringify(this._store));
     this._store = [];
+    this._cleanUp();
   }
 
   public get store(): ITelemetryData[] {
     return this._store;
+  }
+
+  /**
+   * When there are too many log files, delete the old ones.
+   */
+  private _cleanUp(): void {
+    if (fsx.existsSync(this._dataFolder)) {
+      const files: string[] = fsx.readdirSync(this._dataFolder);
+      if (files.length > MAX_FILE_COUNT) {
+        const sortedFiles: string[] = files.map(fileName => {
+          const filePath: string = path.join(this._dataFolder, fileName);
+          const stats: fsx.Stats = fsx.statSync(filePath);
+          return {
+            filePath: filePath,
+            modifiedTime: stats.mtime.getTime(),
+            isFile: stats.isFile()
+          };
+        })
+        .filter(value => {
+          // Only delete files
+          return value.isFile;
+        })
+        .sort((a, b) => {
+          return a.modifiedTime - b.modifiedTime;
+        })
+        .map(s => {
+          return s.filePath;
+        });
+        const filesToDelete: number = sortedFiles.length - MAX_FILE_COUNT;
+        for (let i: number = 0; i < filesToDelete; i++) {
+          fsx.unlinkSync(sortedFiles[i]);
+        }
+      }
+    }
   }
 
   private _getFilePath(): string {