+++ libraries/api-documenter/src/yaml/YamlGenerator.ts
@@ -5,18 +5,34 @@ import * as fsx from 'fs-extra';
 import * as path from 'path';
 import yaml = require('js-yaml');
 import { JsonFile, JsonSchema } from '@microsoft/node-core-library';
-import { MarkupElement, IDocElement } from '@microsoft/api-extractor';
+import {
+  MarkupElement,
+  IDocElement,
+  IApiMethod,
+  IApiConstructor,
+  IApiParameter,
+  IApiProperty,
+  IApiEnumMember,
+  IApiClass,
+  IApiInterface
+} from '@microsoft/api-extractor';
 
 import { DocItemSet, DocItem, DocItemKind, IDocItemSetResolveResult } from '../DocItemSet';
 import {
-  IYamlFile,
-  IYamlItem
-} from './IYamlFile';
+  IYamlApiFile,
+  IYamlItem,
+  IYamlSyntax,
+  IYamlParameter
+} from './IYamlApiFile';
+import {
+  IYamlTocFile,
+  IYamlTocItem
+} from './IYamlTocFile';
 import { RenderingHelpers } from '../RenderingHelpers';
 import { MarkupBuilder } from '../MarkupBuilder';
 import { MarkdownRenderer, IMarkdownRenderApiLinkArgs } from '../MarkdownRenderer';
 
-const yamlSchema: JsonSchema = JsonSchema.fromFile(path.join(__dirname, 'typescript.schema.json'));
+const yamlApiSchema: JsonSchema = JsonSchema.fromFile(path.join(__dirname, 'typescript.schema.json'));
 
 /**
  * Writes documentation in the Universal Reference YAML file format, as defined by typescript.schema.json.
@@ -38,9 +54,11 @@ export class YamlGenerator {
     for (const docPackage of this._docItemSet.docPackages) {
       this._visitDocItems(docPackage, undefined);
     }
+
+    this._writeTocFile(this._docItemSet.docPackages);
   }
 
-  private _visitDocItems(docItem: DocItem, parentYamlFile: IYamlFile | undefined): boolean {
+  private _visitDocItems(docItem: DocItem, parentYamlFile: IYamlApiFile | undefined): boolean {
     const yamlItem: IYamlItem | undefined = this._generateYamlItem(docItem);
     if (!yamlItem) {
       return false;
@@ -52,7 +70,7 @@ export class YamlGenerator {
       }
       parentYamlFile.items.push(yamlItem);
     } else {
-      const newYamlFile: IYamlFile = {
+      const newYamlFile: IYamlApiFile = {
         items: []
       };
       newYamlFile.items.push(yamlItem);
@@ -66,7 +84,13 @@ export class YamlGenerator {
         }
       }
 
-      this._writeYamlFile(newYamlFile, docItem);
+      const yamlFilePath: string = this._getYamlFilePath(docItem);
+
+      if (docItem.kind === DocItemKind.Package) {
+        console.log('Writing ' + this._getYamlFilePath(docItem));
+      }
+
+      this._writeYamlFile(newYamlFile, yamlFilePath, 'UniversalReference', yamlApiSchema);
 
       if (parentYamlFile) {
         if (!parentYamlFile.references) {
@@ -84,6 +108,59 @@ export class YamlGenerator {
     return true;
   }
 
+  /**
+   * Write the table of contents
+   */
+  private _writeTocFile(docItems: DocItem[]): void {
+    const tocFile: IYamlTocFile = {
+      items: [ ]
+    };
+
+    tocFile.items.push({
+      name: 'SharePoint Framework', // TODO: parameterize this
+      items: this._buildTocItems(docItems.filter(x => x.isExternalPackage))
+    });
+
+    const externalPackages: DocItem[] = docItems.filter(x => !x.isExternalPackage);
+    if (externalPackages.length) {
+      tocFile.items.push({
+        name: '─────────────'
+      });
+      tocFile.items.push({
+        name: 'External Packages',
+        items: this._buildTocItems(externalPackages)
+      });
+    }
+
+    this._buildTocItems(docItems);
+    const tocFilePath: string = path.join(this._outputFolder, 'toc.yml');
+    console.log('Writing ' + tocFilePath);
+    this._writeYamlFile(tocFile, tocFilePath, '', undefined);
+  }
+
+  private _buildTocItems(docItems: DocItem[]): IYamlTocItem[] {
+    const tocItems: IYamlTocItem[] = [];
+    for (const docItem of docItems) {
+      if (this._shouldEmbed(docItem.kind)) {
+        // Don't generate table of contents items for embedded definitions
+        continue;
+      }
+
+      const tocItem: IYamlTocItem = {
+        name: RenderingHelpers.getUnscopedPackageName(docItem.name),
+        uid: this._getUid(docItem)
+      };
+
+      tocItems.push(tocItem);
+
+      const childItems: IYamlTocItem[] = this._buildTocItems(docItem.children);
+      if (childItems.length > 0) {
+        tocItem.items = childItems;
+      }
+    }
+    return tocItems;
+  }
+
   private _shouldEmbed(docItemKind: DocItemKind): boolean {
     switch (docItemKind) {
       case DocItemKind.Class:
@@ -99,18 +176,44 @@ export class YamlGenerator {
     const yamlItem: Partial<IYamlItem> = { };
     yamlItem.uid = this._getUid(docItem);
 
-    const summary: string = this._renderMarkdownFromDocElement(docItem.apiItem.summary, docItem);
+    let summary: string = this._renderMarkdownFromDocElement(docItem.apiItem.summary, docItem);
+    const remarks: string = this._renderMarkdownFromDocElement(docItem.apiItem.remarks, docItem);
+
+    if ((docItem.apiItem.deprecatedMessage || []).length > 0) {
+      const deprecatedMessage: string = this._renderMarkdownFromDocElement(docItem.apiItem.deprecatedMessage, docItem);
+
+      summary = '> [!NOTE]\n'
+        + '> _This API is now obsolete._\n'
+        + '> \n'
+        + `> _${deprecatedMessage.trim()}_\n\n`
+        + summary;
+    }
+
+    if (docItem.apiItem.isBeta) {
+      summary = '> [!NOTE]\n'
+        + '> _This API is provided as a preview for developers and may change based on feedback_\n'
+        + '> _that we receive. Do not use this API in a production environment._\n\n'
+        + summary;
+    }
+
+    if (remarks && this._shouldEmbed(docItem.kind)) {
+      // This is a temporary workaround, since "Remarks" are not currently being displayed for embedded items
+      if (summary) {
+        summary += '\n\n';
+      }
+      summary += '### Remarks\n\n' + remarks;
+    }
+
     if (summary) {
       yamlItem.summary = summary;
     }
 
-    const remarks: string = this._renderMarkdownFromDocElement(docItem.apiItem.remarks, docItem);
     if (remarks) {
       yamlItem.remarks = remarks;
     }
 
     yamlItem.name = docItem.name;
-    yamlItem.fullName = docItem.name;
+    yamlItem.fullName = yamlItem.uid;
     yamlItem.langs = [ 'typeScript' ];
 
     switch (docItem.kind) {
@@ -120,23 +223,42 @@ export class YamlGenerator {
       case DocItemKind.Enum:
         yamlItem.type = 'enum';
         break;
+      case DocItemKind.EnumMember:
+        yamlItem.type = 'field';
+        const enumMember: IApiEnumMember = docItem.apiItem as IApiEnumMember;
+        if (enumMember.value) {
+          // NOTE: In TypeScript, enum members can be strings or integers.
+          // If it is an integer, then enumMember.value will be a string representation of the integer.
+          // If it is a string, then enumMember.value will include the quotation marks.
+          // Enum values can also be calculated numbers, however this is not implemented yet.
+          yamlItem.numericValue = enumMember.value as any; // tslint:disable-line:no-any
+        }
+        break;
       case DocItemKind.Class:
         yamlItem.type = 'class';
+        this._populateYamlClassOrInterface(yamlItem, docItem);
         break;
       case DocItemKind.Interface:
         yamlItem.type = 'interface';
+        this._populateYamlClassOrInterface(yamlItem, docItem);
         break;
       case DocItemKind.Method:
         yamlItem.type = 'method';
+        this._populateYamlMethod(yamlItem, docItem);
         break;
       case DocItemKind.Constructor:
         yamlItem.type = 'constructor';
+        this._populateYamlMethod(yamlItem, docItem);
         break;
       case DocItemKind.Property:
         yamlItem.type = 'property';
+        this._populateYamlProperty(yamlItem, docItem);
+        break;
+      case DocItemKind.Function:
+        // Unimplemented
         break;
       default:
-        return undefined;
+        throw new Error('Unimplemented item kind: ' + DocItemKind[docItem.kind as DocItemKind]);
     }
 
     if (docItem.kind !== DocItemKind.Package && !this._shouldEmbed(docItem.kind)) {
@@ -144,6 +266,72 @@ export class YamlGenerator {
     }
 
     return yamlItem as IYamlItem;
+  }
+
+  private _populateYamlClassOrInterface(yamlItem: Partial<IYamlItem>, docItem: DocItem): void {
+    const apiStructure: IApiClass | IApiInterface = docItem.apiItem as IApiClass | IApiInterface;
+
+    let text: string = '';
+
+    if (apiStructure.extends) {
+      text += `**Extends:** \`${apiStructure.extends}\`\n\n`;
+    }
+    if (apiStructure.implements) {
+      text += `**Implements:** \`${apiStructure.implements}\`\n\n`;
+    }
+
+    if (text) {
+      yamlItem.remarks = text + (yamlItem.remarks || '');
+    }
+  }
+
+  private _populateYamlMethod(yamlItem: Partial<IYamlItem>, docItem: DocItem): void {
+    const apiMethod: IApiMethod | IApiConstructor = docItem.apiItem as IApiMethod;
+    yamlItem.name = RenderingHelpers.getConciseSignature(docItem.name, apiMethod);
+
+    const syntax: IYamlSyntax = {
+      content: apiMethod.signature
+    };
+    yamlItem.syntax = syntax;
+
+    if (apiMethod.returnValue) {
+      syntax.return = {
+        type: [ apiMethod.returnValue.type ],
+        description: this._renderMarkdownFromDocElement(apiMethod.returnValue.description, docItem)
+      };
+    }
+
+    const parameters: IYamlParameter[] = [];
+    for (const parameterName of Object.keys(apiMethod.parameters)) {
+      const apiParameter: IApiParameter = apiMethod.parameters[parameterName];
+      parameters.push(
+        {
+           id: parameterName,
+           description:  this._renderMarkdownFromDocElement(apiParameter.description, docItem),
+           type: [ apiParameter.type || '' ]
+        } as IYamlParameter
+      );
+    }
+
+    if (parameters.length) {
+      syntax.parameters = parameters;
+    }
+
+  }
+
+  private _populateYamlProperty(yamlItem: Partial<IYamlItem>, docItem: DocItem): void {
+    const apiProperty: IApiProperty = docItem.apiItem as IApiProperty;
+
+    const syntax: IYamlSyntax = {
+      content: docItem.name + ': ' + apiProperty.type + ';' // TODO
+    };
+    yamlItem.syntax = syntax;
+
+    if (apiProperty.type) {
+      syntax.return = {
+        type: [ apiProperty.type ]
+      };
+    }
   }
 
   private _renderMarkdownFromDocElement(docElements: IDocElement[] | undefined, containingDocItem: DocItem): string {
@@ -162,57 +350,34 @@ export class YamlGenerator {
           // Eventually we should introduce a warnings file
           console.error('==> UNRESOLVED REFERENCE: ' + JSON.stringify(args.reference));
         } else {
-          // We will calculate a relativeUrl to the nearest non-embedded DocItem.
-          // (The rest will be determined by the URL fragment.)
-          let nonEmbeddedDocItem: DocItem = result.docItem;
-          while (this._shouldEmbed(nonEmbeddedDocItem.kind) && nonEmbeddedDocItem.parent) {
-            nonEmbeddedDocItem = nonEmbeddedDocItem.parent;
-          }
-
-          const currentFolder: string = path.dirname(this._getYamlFilePath(containingDocItem));
-          const targetFilePath: string = this._getYamlFilePath(nonEmbeddedDocItem);
-          const relativePath: string = path.relative(currentFolder, targetFilePath);
-          let relativeUrl: string = relativePath
-            .replace(/[\\]/g, '/') // replace all backslashes with slashes
-            .replace(/\.[^\.\\/]+$/, ''); // remove file extension
-
-          if (relativeUrl.substr(0, 1) !== '.') {
-            // If the path doesn't already start with "./", then add this prefix.
-            relativeUrl = './' + relativeUrl;
-          }
-
-          // Do we need to link to a fragment within the page?
-          if (nonEmbeddedDocItem !== result.docItem) {
-            const urlFragment: string = this._getEmbeddedUrlFragment(result.docItem);
-            args.prefix = '[';
-            args.suffix = `](${relativeUrl}#${urlFragment})`;
-          } else {
-            args.prefix = '[';
-            args.suffix = `](${relativeUrl})`;
-          }
-
+          args.prefix = '[';
+          args.suffix = `](xref:${this._getUid(result.docItem)})`;
         }
       }
     });
   }
 
-  private _writeYamlFile(yamlFile: IYamlFile, docItem: DocItem): void {
-    const yamlFilePath: string = this._getYamlFilePath(docItem);
+  private _writeYamlFile(dataObject: {}, filePath: string, yamlMimeType: string,
+    schema: JsonSchema|undefined): void {
 
-    if (docItem.kind === DocItemKind.Package) {
-      console.log('Writing ' + this._getYamlFilePath(docItem));
-    }
+    JsonFile.validateNoUndefinedMembers(dataObject);
 
-    JsonFile.validateNoUndefinedMembers(yamlFile);
-
-    const stringified: string = '### YamlMime:UniversalReference\n' + yaml.safeDump(yamlFile, {
+    let stringified: string = yaml.safeDump(dataObject, {
       lineWidth: 120
     });
+
+    if (yamlMimeType) {
+      stringified = `### YamlMime:${yamlMimeType}\n` + stringified;
+    }
+
     const normalized: string = stringified.split('\n').join('\r\n');
 
-    fsx.mkdirsSync(path.dirname(yamlFilePath));
-    fsx.writeFileSync(yamlFilePath, normalized);
-    yamlSchema.validateObject(yamlFile, yamlFilePath);
+    fsx.mkdirsSync(path.dirname(filePath));
+    fsx.writeFileSync(filePath, normalized);
+
+    if (schema) {
+      schema.validateObject(dataObject, filePath);
+    }
   }
 
   /**
@@ -233,15 +398,6 @@ export class YamlGenerator {
       }
     }
     return result;
-  }
-
-  /**
-   * Calculate the HTML anchor fragment for DocItems that are embedded in a web page containing
-   * other items.
-   * Example:  node_core_library_JsonFile_load
-   */
-  private _getEmbeddedUrlFragment(docItem: DocItem): string {
-    return this._getUid(docItem).replace(/\./g, '_'); // replace all periods with underscores
   }
 
   private _getYamlFilePath(docItem: DocItem): string {