+++ libraries/api-extractor/src/aedoc/test/ApiDocumentation.test.ts
@@ -29,7 +29,7 @@ const compilerOptions: ts.CompilerOptions = {
   module: ts.ModuleKind.CommonJS,
   moduleResolution: ts.ModuleResolutionKind.NodeJs,
   rootDir: inputFolder,
-  typeRoots: ['./'] // We need to ignore @types in these tests
+  typeRoots: [] // We need to ignore @types in these tests
 };
 const extractor: Extractor = new Extractor({
   compilerOptions: compilerOptions,