+++ libraries/api-extractor/src/generators/test/ApiFileGenerator.test.ts
@@ -51,7 +51,7 @@ describe('ApiFileGenerator tests', function (): void {
         module: ts.ModuleKind.CommonJS,
         moduleResolution: ts.ModuleResolutionKind.NodeJs,
         rootDir: inputFolder,
-        typeRoots: ['./'] // We need to ignore @types in these tests
+        typeRoots: [] // We need to ignore @types in these tests
       };
       const extractor: Extractor = new Extractor({
         compilerOptions: compilerOptions,
@@ -115,7 +115,7 @@ describe('ApiFileGenerator tests', function (): void {
         module: ts.ModuleKind.CommonJS,
         moduleResolution: ts.ModuleResolutionKind.NodeJs,
         rootDir: inputFolder,
-        typeRoots: ['./'] // We need to ignore @types in these tests
+        typeRoots: [] // We need to ignore @types in these tests
       };
       const extractor: Extractor = new Extractor({
         compilerOptions: compilerOptions,