+++ libraries/api-extractor/CHANGELOG.md
@@ -1,6 +1,40 @@
 # Change Log - @microsoft/api-extractor
 
-This log was last generated on Thu, 07 Sep 2017 00:11:11 GMT and should not be manually modified.
+This log was last generated on Fri, 22 Sep 2017 01:04:02 GMT and should not be manually modified.
+
+## 3.3.0
+Fri, 22 Sep 2017 01:04:02 GMT
+
+### Minor changes
+
+- Upgrade to es6
+
+## 3.2.6
+Wed, 20 Sep 2017 22:10:17 GMT
+
+*Changes not tracked*
+
+## 3.2.5
+Mon, 11 Sep 2017 13:04:55 GMT
+
+### Patches
+
+- The isBeta and deprecatedMessage fields are now inherited in the *.api.json files
+- Fix an issue where the *.api.json file was sometimes missing function parameters
+
+## 3.2.4
+Fri, 08 Sep 2017 01:28:04 GMT
+
+### Patches
+
+- Deprecate @types/es6-coll ections in favor of built-in typescript typings 'es2015.collection' a nd 'es2015.iterable'
+
+## 3.2.3
+Thu, 07 Sep 2017 13:04:35 GMT
+
+### Patches
+
+- Fix incorrect schema/typings for enum members
 
 ## 3.2.2
 Thu, 07 Sep 2017 00:11:11 GMT