+++ libraries/node-core-library/CHANGELOG.md
@@ -1,6 +1,42 @@
 # Change Log - @microsoft/node-core-library
 
-This log was last generated on Thu, 07 Sep 2017 00:11:12 GMT and should not be manually modified.
+This log was last generated on Fri, 22 Sep 2017 01:04:02 GMT and should not be manually modified.
+
+## 0.3.0
+Fri, 22 Sep 2017 01:04:02 GMT
+
+### Minor changes
+
+- Upgrade to es6
+
+## 0.2.11
+Wed, 20 Sep 2017 22:10:17 GMT
+
+*Changes not tracked*
+
+## 0.2.10
+Mon, 11 Sep 2017 13:04:55 GMT
+
+*Changes not tracked*
+
+## 0.2.9
+Fri, 08 Sep 2017 13:04:00 GMT
+
+### Patches
+
+- Improve error reporting for JsonFile.validateNoUndefinedMembers()
+
+## 0.2.8
+Fri, 08 Sep 2017 01:28:04 GMT
+
+### Patches
+
+- Deprecate @types/es6-coll ections in favor of built-in typescript typings 'es2015.collection' a nd 'es2015.iterable'
+
+## 0.2.7
+Thu, 07 Sep 2017 13:04:35 GMT
+
+*Changes not tracked*
 
 ## 0.2.6
 Thu, 07 Sep 2017 00:11:12 GMT